Backend/requirements.txt
uvicorn==0.25.0
fastapi==0.106.0
pydantic==2.5.3
pandas==2.1.4
scikit-learn == 1.3.2
openai==1.6.1
langchain==0.0.353
python-dotenv==1.0.0
